fix: address code review feedback for metrics system
Agent-Logs-Url: https://github.com/EntglDb/BLite/sessions/7a6a6b96-270b-4cb4-98b5-c9a31dbb2a9e

+++ src/BLite.Core/Metrics/MetricsDispatcher.cs
@@ -160,7 +160,7 @@ internal sealed class MetricsDispatcher : IDisposable
 
     private void Aggregate(MetricEvent evt)
     {
-        long micros = (long)(evt.ElapsedMicros); // already in µs
+        long micros = evt.ElapsedMicros;
 
         switch (evt.Type)
         {
